magento-engcom/bulk-api#4 Support for Async operations in WebAPI
- Reduced coupling between objects
